Update Struct `StakingCondition`
* uint256 timeUnit => uint80 timeUnit
* uint256 startTimeStamp => uint80 startTimeStamp
* uint256 endTimeStamp => uint80 endTimeStamp

These changes reduce storage slot usage for this struct from 4 to 2.

+++ contracts/base/Staking721Base.sol
@@ -55,7 +55,7 @@ contract Staking721Base is ContractMetadata, Multicall, Ownable, Staking721, ERC
     uint256 private rewardTokenBalance;
 
     constructor(
-        uint256 _timeUnit,
+        uint80 _timeUnit,
         uint256 _rewardsPerUnitTime,
         address _stakingToken,
         address _rewardToken,
